Committer: pi <pi@nuc.lan.local>  On branch main  Your branch is up to date with 'origin/main'.  Changes to be committed: 	modified:   bgtech_meter_registers.h

+++ esphome/components/bgtech_meter/bgtech_meter_registers.h
@@ -10,23 +10,23 @@ static const float MULTIPLY_TEN_UNIT = 10;
 static const float MULTIPLY_THOUSAND_UNIT = 1000;
 
 /* PHASE STATUS REGISTERS */
-static const uint32_t BGTECH_TOTAL_ACTIVE_ENERGY = 0x0122;
-static const uint32_t BGTECH_IMPORT_ACTIVE_ENERGY = 0x010E;
-static const uint32_t BGTECH_EXPORT_ACTIVE_ENERGY = 0x0118;
-static const uint32_t BGTECH_TOTAL_REACTIVE_ENERGY = 0x0140;
-static const uint32_t BGTECH_IMPORT_REACTIVE_ENERGY = 0x012C;
-static const uint32_t BGTECH_EXPORT_REACTIVE_ENERGY = 0x0136;
-static const uint32_t BGTECH_APPARENT_ENERGY = 0x0175;
-static const uint32_t BGTECH_ACTIVE_POWER = 0x0104;
-static const uint32_t BGTECH_REACTIVE_POWER = 0x0108;
-static const uint32_t BGTECH_APPARENT_POWER = 0x0106;
-static const uint32_t BGTECH_VOLTAGE = 0x0100;
-static const uint32_t BGTECH_CURRENT = 0x0102;
-static const uint16_t BGTECH_POWER_FACTOR = 0x010B;
-static const uint16_t BGTECH_FREQUENCY = 0x010A;
-static const uint32_t BGTECH_MAXIMUM_DEMAND_ACTIVE_POWER = 0x0178;
-static const uint32_t BGTECH_MAXIMUM_DEMAND_REACTIVE_POWER = 0x0182;
-static const uint16_t BGTECH_MAXIMUM_DEMAND_APPARENT_POWER = 0x0180;
+static const uint16_t BGTECH_TOTAL_ACTIVE_ENERGY = 0x0122;
+static const uint16_t BGTECH_IMPORT_ACTIVE_ENERGY = 0x010E;
+static const uint16_t BGTECH_EXPORT_ACTIVE_ENERGY = 0x0118;
+static const uint16_t BGTECH_TOTAL_REACTIVE_ENERGY = 0x0140;
+static const uint16_t BGTECH_IMPORT_REACTIVE_ENERGY = 0x012C;
+static const uint16_t BGTECH_EXPORT_REACTIVE_ENERGY = 0x0136;
+static const uint16_t BGTECH_APPARENT_ENERGY = 0x0175;
+static const uint16_t BGTECH_ACTIVE_POWER = 0x0104;
+static const uint16_t BGTECH_REACTIVE_POWER = 0x0108;
+static const uint16_t BGTECH_APPARENT_POWER = 0x0106;
+static const uint16_t BGTECH_VOLTAGE = 0x0100;
+static const uint16_t BGTECH_CURRENT = 0x0102;
+static const uint8_t BGTECH_POWER_FACTOR = 0x010B;
+static const uint8_t BGTECH_FREQUENCY = 0x010A;
+static const uint16_t BGTECH_MAXIMUM_DEMAND_ACTIVE_POWER = 0x0178;
+static const uint16_t BGTECH_MAXIMUM_DEMAND_REACTIVE_POWER = 0x0182;
+static const uint8_t BGTECH_MAXIMUM_DEMAND_APPARENT_POWER = 0x0180;
 
 }  // namespace bgtech_meter
 }  // namespace esphome
